requirements.txt
streamlit==1.11.0
streamlit-option-menu==0.3.2
pandas==1.3.4
numpy==1.20.3
tensorflow==2.9.1
transformers==4.21.2
plotly-express==0.4.1
protobuf~=3.19.0
openpyxl==3.0.9